requirements.txt
djangorestframework-jwt
beautifulsoup4==4.8.2
dj-database-url==0.5.0
Django==3.0.4
django-bootstrap4==1.1.1
django-crispy-forms==1.9.0
django-taggit==1.2.0
djangorestframework==3.11.0
et-xmlfile==1.0.1
gunicorn==20.0.4
jdcal==1.4.1
openpyxl==3.0.3
Pillow==7.0.0
psycopg2==2.8.4
pytz==2019.3
soupsieve==2.0
whitenoise==5.0.1
xlwt==1.3.0
